split workflow and publishing
