+++ Kontent.Ai.Management/ManagementClient.Workflow.cs
@@ -1,5 +1,4 @@
-﻿using Kontent.Ai.Management.Models.LanguageVariants;
-using Kontent.Ai.Management.Models.Shared;
+﻿using Kontent.Ai.Management.Models.Shared;
 using Kontent.Ai.Management.Models.Workflow;
 using System;
 using System.Collections.Generic;
@@ -59,124 +58,5 @@ public sealed partial class ManagementClient
 
         var endpointUrl = _urlBuilder.BuildWorkflowsUrl(identifier);
         await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Delete);
-    }
-
-    /// <inheritdoc />
-    public async Task ChangeLanguageVariantWorkflowAsync(LanguageVariantIdentifier identifier, WorkflowStepIdentifier workflowStepIdentifier)
-    {
-        if (identifier == null)
-        {
-            throw new ArgumentNullException(nameof(identifier));
-        }
-
-        if (workflowStepIdentifier == null)
-        {
-            throw new ArgumentNullException(nameof(workflowStepIdentifier));
-        }
-
-        var endpointUrl = _urlBuilder.BuildWorkflowChangeUrl(identifier);
-
-        await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Put, workflowStepIdentifier);
-    }
-
-    /// <inheritdoc />
-    public async Task PublishLanguageVariantAsync(LanguageVariantIdentifier identifier)
-    {
-        if (identifier == null)
-        {
-            throw new ArgumentNullException(nameof(identifier));
-        }
-
-        var endpointUrl = _urlBuilder.BuildPublishVariantUrl(identifier);
-
-        await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Put);
-    }
-
-    /// <inheritdoc />
-    public async Task SchedulePublishingOfLanguageVariantAsync(LanguageVariantIdentifier identifier, ScheduleModel scheduleModel)
-    {
-        if (identifier == null)
-        {
-            throw new ArgumentNullException(nameof(identifier));
-        }
-
-        if (scheduleModel == null)
-        {
-            throw new ArgumentNullException(nameof(scheduleModel));
-        }
-
-        var endpointUrl = _urlBuilder.BuildPublishVariantUrl(identifier);
-
-        await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Put, scheduleModel);
-    }
-
-    /// <inheritdoc />
-    public async Task CancelPublishingOfLanguageVariantAsync(LanguageVariantIdentifier identifier)
-    {
-        if (identifier == null)
-        {
-            throw new ArgumentNullException(nameof(identifier));
-        }
-
-        var endpointUrl = _urlBuilder.BuildCancelPublishingVariantUrl(identifier);
-
-        await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Put);
-    }
-
-    /// <inheritdoc />
-    public async Task UnpublishLanguageVariantAsync(LanguageVariantIdentifier identifier)
-    {
-        if (identifier == null)
-        {
-            throw new ArgumentNullException(nameof(identifier));
-        }
-
-        var endpointUrl = _urlBuilder.BuildUnpublishVariantUrl(identifier);
-
-        await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Put);
-    }
-
-    /// <inheritdoc />
-    public async Task CancelUnpublishingOfLanguageVariantAsync(LanguageVariantIdentifier identifier)
-    {
-        if (identifier == null)
-        {
-            throw new ArgumentNullException(nameof(identifier));
-        }
-
-        var endpointUrl = _urlBuilder.BuildCancelUnpublishingVariantUrl(identifier);
-
-        await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Put);
-    }
-
-    /// <inheritdoc />
-    public async Task ScheduleUnpublishingOfLanguageVariantAsync(LanguageVariantIdentifier identifier, ScheduleModel scheduleModel)
-    {
-        if (identifier == null)
-        {
-            throw new ArgumentNullException(nameof(identifier));
-        }
-
-        if (scheduleModel == null)
-        {
-            throw new ArgumentNullException(nameof(scheduleModel));
-        }
-
-        var endpointUrl = _urlBuilder.BuildUnpublishVariantUrl(identifier);
-
-        await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Put, scheduleModel);
-    }
-
-    /// <inheritdoc />
-    public async Task CreateNewVersionOfLanguageVariantAsync(LanguageVariantIdentifier identifier)
-    {
-        if (identifier == null)
-        {
-            throw new ArgumentNullException(nameof(identifier));
-        }
-
-        var endpointUrl = _urlBuilder.BuildNewVersionVariantUrl(identifier);
-
-        await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Put);
     }
 }
